Refactored code to better suit rebase.
As bitmaps now always have an allocated surface, there is no need to
ensure the surface exists.

Surfaces are now also always drawable (which means that
bitmap_be.drawable is probably no longer needed. This was done because
it seemed the only reason surfaces were not drawable because of images
that had been loaded from file which needed to be kept in sync with
their textures. My modifications now no longer make this a problem as
data will always be synced between surfaces when different drawing
operations call for it.